IDS/requirements.txt
opencv-python
torch
ultralytics
deep-oc-sort
mediapipe
pydantic